+++ src/sync/mod.rs
@@ -73,4 +73,4 @@ pub use mutex::{Mutex, MutexGuard};
 pub use notify::Notify;
 pub use once_cell::OnceCell;
 pub use rw_lock::{RwLock, RwLockReadFuture, RwLockReadGuard, RwLockWriteFuture, RwLockWriteGuard};
-pub use semaphore::{Permit, Semaphore};
+pub use semaphore::{Semaphore, SemaphorePermit};